add memcov eval against tzer

+++ experiments/cov_eval.py
@@ -13,11 +13,11 @@ import os
 import sys
 import random
 from time import time
-import numpy as np
 import argparse
 import subprocess
 
 from tqdm import tqdm
+import numpy as np
 
 from nnsmith.util import mkdir
 
@@ -50,7 +50,8 @@ if __name__ == '__main__':
     parser.add_argument('--resume', action='store_true', help='Resume eval.')
     parser.add_argument('--seed', type=int, default=233,
                         help='to generate random input data')
-    parser.add_argument('--lib', type=str, required=True,
+    parser.add_argument('--memcov', action='store_true', help='Use memcov.')
+    parser.add_argument('--lib', type=str, default=None,
                         help='path to instrumented library')
     parser.add_argument('--max_time', type=int, default=60 *
                         60 * 4, help='max time in seconds for coverage evaluation')
@@ -64,14 +65,22 @@ if __name__ == '__main__':
     np.random.seed(args.seed)
 
     HAS_LZ4 = os.system('lz4 --version > /dev/null 2>&1') == 0
-    if not HAS_LZ4:
+    if not HAS_LZ4 and not args.memcov:
         print('==> lz4 not found. Storing lcov w/o compression is disk-killing. Please install lz4!')
         exit(1)
 
-    lib_expr = ''
-    for lib in args.lib.split():
-        assert os.path.exists(lib), f'{lib} does not exist!'
-        lib_expr += f' -object {os.path.realpath(lib)} '
+    if args.lib:
+        lib_expr = ''
+        for lib in args.lib.split():
+            assert os.path.exists(lib), f'{lib} does not exist!'
+            lib_expr += f' -object {os.path.realpath(lib)} '
+    else:
+        assert args.memcov, 'you need to provide either --lib for ast cov or --memcov for memcov.'
+
+    if args.memcov and args.dp < 1000:
+        print('==> For memcov, you need at least 1k data points as we lose cov when it crashes.')
+        print(f'==> Setting data point from {args.dp} to 1000.')
+        args.dp = 1000
 
     if not args.resume:
         mkdir(args.report_folder)
@@ -104,8 +113,8 @@ if __name__ == '__main__':
         config_file = open(os.path.join(args.report_folder, 'stats.csv'), 'a')
         stderr_file = open(os.path.join(args.report_folder, 'stderr.txt'), 'a')
 
-    def record(btime, n_models, seed, lcov_name):
-        config_file.write(f'{btime},{seed},{n_models},{lcov_name}\n')
+    def record(btime, n_models, seed, cov_name):
+        config_file.write(f'{btime},{seed},{n_models},{cov_name}\n')
         config_file.flush()
 
     lines = open(os.path.join(args.model_dir, 'gentime.csv'), 'r').readlines()
@@ -133,8 +142,15 @@ if __name__ == '__main__':
             batch = batch_list[i]
             btime = 0
             sum_path = os.path.join(args.report_folder, f'{i}.txt')
+
+            # for source cov;
             lcov_name = f'{i}.lcov'
             lcov_path = os.path.join(args.report_folder, lcov_name)
+
+            # for memcov
+            memcov_name = f'{i}.memcov'
+            memcov_path = os.path.join(args.report_folder, memcov_name)
+            
             seed = random.getrandbits(32)
 
             model_batch = []
@@ -152,20 +168,26 @@ if __name__ == '__main__':
             copied_env["LLVM_PROFILE_FILE"] = str(profraw_path)
 
             tstart = time()          # <=== START
-            p = subprocess.Popen([
+            arguments = [
                 'python', 'experiments/batch_eval.py',
                 '--models', *model_batch,
                 '--backend', args.backend,
                 '--dev', args.dev,
                 '--seed', str(seed),
-            ], stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=copied_env)
+            ]
+            if args.memcov:
+                arguments += ['--memcov', memcov_path]
+            p = subprocess.Popen(arguments, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=copied_env)
             _, errs = p.communicate()
             errs = errs.decode()
             exit_code = p.returncode
 
+            if exit_code != 0:
+                print(f'==> Batch {i} process crashed!')
+
             # Write stderr
             stderr_file.write(
-                f'iter {i}: {[os.path.split(path)[-1] for path in model_batch]} ~ exit_code={exit_code}\n')
+                f'iter {i}: =================> EXIT CODE {exit_code}\n')
             if errs:
                 stderr_file.write(errs)
             stderr_file.flush()
@@ -181,36 +203,42 @@ if __name__ == '__main__':
             # Wrap up this batch.
 
             # Get coverage report
-            if os.path.exists(profraw_path):
-                llvm_profdata = 'llvm-profdata'
-                llvm_cov = 'llvm-cov'
-                if args.llvm_version and str(args.llvm_version).isnumeric():
-                    llvm_profdata += f'-{args.llvm_version}'
-                    llvm_cov += f'-{args.llvm_version}'
+            if not args.memcov:
+                if os.path.exists(profraw_path):
+                    llvm_profdata = 'llvm-profdata'
+                    llvm_cov = 'llvm-cov'
+                    if args.llvm_version and str(args.llvm_version).isnumeric():
+                        llvm_profdata += f'-{args.llvm_version}'
+                        llvm_cov += f'-{args.llvm_version}'
 
-                profdata_path = os.path.join(
-                    args.report_folder, f'{i}.profdata')
-                # summary might be useless as it does not consider prior runs.
-                if 0 != os.system(f'{llvm_profdata} merge -sparse {profraw_path} -o {profdata_path}') or \
-                        0 != os.system(f'{llvm_cov} export -instr-profile={profdata_path} -format=lcov {lib_expr} > {lcov_path}'):
-                    print(f'Getting coverage failed!!', file=sys.stderr)
-                else:  # clean temporary files
-                    if args.sum:
-                        os.system(
-                            f'{llvm_cov} report -instr-profile={profdata_path} {lib_expr} > {sum_path}')
-                    assert 0 == os.system(f'lz4 {lcov_path}')
-                    os.remove(profraw_path)
-                    os.remove(profdata_path)
-            else:
-                print(f'{profraw_path} does not exist...', file=sys.stderr)
+                    profdata_path = os.path.join(
+                        args.report_folder, f'{i}.profdata')
+                    # summary might be useless as it does not consider prior runs.
+                    if 0 != os.system(f'{llvm_profdata} merge -sparse {profraw_path} -o {profdata_path}') or \
+                            0 != os.system(f'{llvm_cov} export -instr-profile={profdata_path} -format=lcov {lib_expr} > {lcov_path}'):
+                        print(f'Getting coverage failed!!', file=sys.stderr)
+                    else:  # clean temporary files
+                        if args.sum:
+                            os.system(
+                                f'{llvm_cov} report -instr-profile={profdata_path} {lib_expr} > {sum_path}')
+                        assert 0 == os.system(f'lz4 {lcov_path}')
+                        os.remove(profraw_path)
+                        os.remove(profdata_path)
+                else:
+                    print(f'{profraw_path} does not exist...', file=sys.stderr)
 
             if os.path.exists(lcov_path + '.lz4'):
                 record(btime + lagged_time, len(model_batch) +
-                       lagged_n_model, seed, lcov_name=lcov_name)
+                    lagged_n_model, seed, cov_name=lcov_name)
+                lagged_time = 0
+                lagged_n_model = 0
+            elif os.path.exists(memcov_path + '.pkl'):
+                record(btime + lagged_time, len(model_batch) +
+                    lagged_n_model, seed, cov_name=memcov_name)
                 lagged_time = 0
                 lagged_n_model = 0
             else:
-                # Means no lcov due to some LLVM issues.
+                # Means no lcov or memcov due to some LLVM issues.
                 lagged_time += btime
                 lagged_n_model += len(model_batch)
